feat(dtfs2-6892): add controllers and error handler wrapper to start routes
