+++ portal/server/controllers/start/index.js
@@ -1,0 +1,5 @@
+const beforeYouStartController = require('./before-you-start.controller');
+const beforeYouStartBankDealController = require('./before-you-start-bank-deal.controller');
+const unableToProceedController = require('./unable-to-proceed.controller');
+
+export { beforeYouStartController, beforeYouStartBankDealController, unableToProceedController };
