+++ portal/server/controllers/start/unable-to-proceed.controller.js
@@ -1,0 +1,3 @@
+export const getUnableToProceedPage = (req, res) => {
+  res.render('unable-to-proceed.njk', { user: req.session.user });
+};
